[BUGFIX] Ensure TypoScript settings are available for cached pages
When a frontend page is cached, TypoScriptFrontendController does
not parse the TypoScript settings, so they can not be checked e.g.
when a user must change the password.

This change ensures, that TypoScript is parsed when needed (only
when user must change password or when exclude PID check is required)

Refs #18